rfac: Cleanup testutls
