+++ internal/testutil/testutil.go
@@ -16,10 +16,7 @@ package testutil
 
 import (
 	"encoding/json"
-	"io"
 	"net/http/httptest"
-	"os"
-	"strings"
 	"testing"
 )
 
@@ -29,53 +26,10 @@ import (
 // ─────────────────────────────────────────────────────────────
 //
 
-func ExpectExitPanic(t *testing.T, code *int, want int) {
-	t.Helper()
-	if r := recover(); r != "exit" {
-		t.Fatalf("expected exit panic, got %v", r)
-	}
-	if *code != want {
-		t.Fatalf("expected exit code %d, got %d", want, *code)
-	}
-}
-
-func StubExit(t *testing.T, target *func(int)) *int {
-	t.Helper()
-	var code int
-	old := *target
-	*target = func(c int) {
-		code = c
-		panic("exit")
-	}
-	t.Cleanup(func() { *target = old })
-	return &code
-}
-
-func CaptureStderr(t *testing.T, f func()) string {
-	t.Helper()
-	old := os.Stderr
-	r, w, _ := os.Pipe()
-	os.Stderr = w
-
-	f()
-
-	w.Close()
-	os.Stderr = old
-	out, _ := io.ReadAll(r)
-	return string(out)
-}
-
 func ExpectStatus(t *testing.T, rec *httptest.ResponseRecorder, want int) {
 	t.Helper()
 	if rec.Code != want {
 		t.Fatalf("expected %d, got %d (%s)", want, rec.Code, rec.Body.String())
-	}
-}
-
-func ExpectBodyContains(t *testing.T, rec *httptest.ResponseRecorder, substr string) {
-	t.Helper()
-	if !strings.Contains(rec.Body.String(), substr) {
-		t.Fatalf("expected body to contain %q, got %q", substr, rec.Body.String())
 	}
 }
 
@@ -87,14 +41,3 @@ func DecodeJSON[T any](t *testing.T, data []byte) T {
 	}
 	return v
 }
-
-func EnvSliceToMap(env []string) map[string]string {
-	out := make(map[string]string, len(env))
-	for _, kv := range env {
-		k, v, _ := strings.Cut(kv, "=")
-		out[k] = v
-	}
-	return out
-}
-
-func Sh(script string) []string { return []string{"sh", "-c", script} }
